src/manage_permissions/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-bigquery==3.12.0